+++ taskiq/__init__.py
@@ -1,7 +1,8 @@
 """Distributed task manager."""
 from taskiq.abc.broker import AsyncBroker, AsyncTaskiqDecoratedTask
-from taskiq.abc.result_backend import AsyncResultBackend, AsyncTaskiqTask
+from taskiq.abc.result_backend import AsyncResultBackend
 from taskiq.message import TaskiqMessage
+from taskiq.task import AsyncTaskiqTask
 
 __all__ = [
     "TaskiqMessage",
