+++ taskiq/abc/broker.py
@@ -1,270 +1,15 @@
 from abc import ABC, abstractmethod
-from dataclasses import asdict, is_dataclass
 from functools import wraps
 from logging import getLogger
-from typing import (  # noqa: WPS235
-    Any,
-    AsyncGenerator,
-    Callable,
-    Coroutine,
-    Dict,
-    Generic,
-    Optional,
-    Set,
-    TypeVar,
-    Union,
-    overload,
-)
-from uuid import uuid4
+from typing import Any, AsyncGenerator, Callable, Dict, Optional, Union, overload
 
-from pydantic import BaseModel
-from typing_extensions import ParamSpec
-
-from taskiq.abc.result_backend import AsyncResultBackend, AsyncTaskiqTask
+from taskiq.abc.result_backend import AsyncResultBackend
+from taskiq.decor import AsyncTaskiqDecoratedTask
 from taskiq.message import TaskiqMessage
 from taskiq.result_backends.dummy import DummyResultBackend
+from taskiq.types_helpers import T_, FuncParams_, ReturnType_
 
 logger = getLogger("taskiq")
-
-_T = TypeVar("_T")  # noqa: WPS111
-_FuncParams = ParamSpec("_FuncParams")
-_ReturnType = TypeVar("_ReturnType")
-
-
-class AsyncKicker(Generic[_FuncParams, _ReturnType]):
-    """Class that used to modify data before sending it to broker."""
-
-    def __init__(
-        self,
-        task_name: str,
-        broker: "AsyncBroker",
-        labels: Dict[str, Any],
-    ) -> None:
-        self.task_name = task_name
-        self.broker = broker
-        self.labels = labels
-
-    def with_label(
-        self,
-        label_name: str,
-        value: Any,
-    ) -> "AsyncKicker[_FuncParams, _ReturnType]":
-        """
-        Update one single label.
-
-        This method is used to update
-        task's labels before sending.
-
-        :param label_name: name of the label to update.
-        :param value: label's value.
-        :return: kicker object with new labels.
-        """
-        self.labels[label_name] = value
-        return self
-
-    def with_labels(
-        self,
-        labels: Dict[str, Any],
-    ) -> "AsyncKicker[_FuncParams, _ReturnType]":
-        """
-        Update function's labels before sending.
-
-        :param labels: dict with new labels.
-        :return: kicker with new labels.
-        """
-        self.labels.update(labels)
-        return self
-
-    def with_broker(
-        self,
-        broker: "AsyncBroker",
-    ) -> "AsyncKicker[_FuncParams, _ReturnType]":
-        """
-        Replace broker for the function.
-
-        This method can be used with
-        shared tasks.
-
-        :param broker: new broker instance.
-        :return: Kicker with new broker.
-        """
-        self.broker = broker
-        return self
-
-    @overload
-    async def kiq(  # noqa: D102
-        self: "AsyncKicker[_FuncParams, Coroutine[Any, Any, _T]]",
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_T]:
-        ...
-
-    @overload
-    async def kiq(  # noqa: D102
-        self: "AsyncKicker[_FuncParams, _ReturnType]",
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_ReturnType]:
-        ...
-
-    async def kiq(
-        self,
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> Any:
-        """
-        This method sends function call over the network.
-
-        It gets current broker and calls it's kick method,
-        returning what it returns.
-
-        :param args: function's arguments.
-        :param kwargs: function's key word arguments.
-
-        :returns: taskiq task.
-        """
-        logger.debug(
-            f"Kicking {self.task_name} with args={args} and kwargs={kwargs}.",
-        )
-        message = self._prepare_message(*args, **kwargs)
-        await self.broker.kick(message)
-        return self.broker.result_backend.generate_task(message.task_id)
-
-    @classmethod
-    def _prepare_arg(cls, arg: Any) -> Any:
-        """
-        Parses argument if possible.
-
-        This function is used to construct dicts
-        from pydantic models or dataclasses.
-
-        :param arg: argument to format.
-        :return: Formatted argument.
-        """
-        if isinstance(arg, BaseModel):
-            arg = arg.dict()
-        if is_dataclass(arg):
-            arg = asdict(arg)
-        return arg
-
-    def _prepare_message(  # noqa: WPS210
-        self,
-        *args: Any,
-        **kwargs: Any,
-    ) -> TaskiqMessage:
-        """
-        Create a message from args and kwargs.
-
-        :param args: function's args.
-        :param kwargs: function's kwargs.
-        :return: constructed message.
-        """
-        formatted_args = []
-        formatted_kwargs = {}
-        for arg in args:
-            formatted_args.append(self._prepare_arg(arg))
-        for kwarg_name, kwarg_val in kwargs.items():
-            formatted_kwargs[kwarg_name] = self._prepare_arg(kwarg_val)
-
-        task_id = uuid4().hex
-
-        return TaskiqMessage(
-            task_id=task_id,
-            task_name=self.task_name,
-            meta=self.labels,
-            args=formatted_args,
-            kwargs=formatted_kwargs,
-        )
-
-
-class AsyncTaskiqDecoratedTask(Generic[_FuncParams, _ReturnType]):
-    """
-    Class for all task functions.
-
-    When function is decorated
-    with the `task` decorator, it
-    will return an instance of this class.
-
-    This class parametrized with original function's
-    arguments types and a return type.
-
-    This class has kiq method which is used
-    to kick tasks out of this thread and send them to
-    current broker.
-    """
-
-    def __init__(
-        self,
-        broker: "AsyncBroker",
-        task_name: str,
-        original_func: Callable[_FuncParams, _ReturnType],
-        labels: Dict[str, Any],
-    ) -> None:
-        self.broker = broker
-        self.task_name = task_name
-        self.original_func = original_func
-        self.labels = labels
-
-    # Docs for this method are ommited in order to help
-    # your IDE resolve correct docs for it.
-    def __call__(  # noqa: D102
-        self,
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> _ReturnType:
-        return self.original_func(*args, **kwargs)
-
-    @overload
-    async def kiq(  # noqa: D102
-        self: "AsyncTaskiqDecoratedTask[_FuncParams, Coroutine[Any, Any, _T]]",
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_T]:
-        ...
-
-    @overload
-    async def kiq(  # noqa: D102
-        self: "AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType]",
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_ReturnType]:
-        ...
-
-    async def kiq(
-        self,
-        *args: _FuncParams.args,
-        **kwargs: _FuncParams.kwargs,
-    ) -> Any:
-        """
-        This method sends function call over the network.
-
-        It gets current broker and calls it's kick method,
-        returning what it returns.
-
-        :param args: function's arguments.
-        :param kwargs: function's key word arguments.
-
-        :returns: taskiq task.
-        """
-        return await self.kicker().kiq(*args, **kwargs)
-
-    def kicker(self) -> AsyncKicker[_FuncParams, _ReturnType]:
-        """
-        This function returns kicker object.
-
-        Kicker is a object that can modyfy kiq request
-        before sendig it.
-
-        :return: AsyncKicker instance.
-        """
-        return AsyncKicker(
-            task_name=self.task_name,
-            broker=self.broker,
-            labels=self.labels,
-        )
-
-    def __repr__(self) -> str:
-        return f"AsyncTaskiqDecoratedTask({self.task_name})"
 
 
 class AsyncBroker(ABC):
@@ -280,7 +25,7 @@ class AsyncBroker(ABC):
 
     def __init__(
         self,
-        result_backend: Optional[AsyncResultBackend[_T]] = None,
+        result_backend: Optional[AsyncResultBackend[T_]] = None,
     ) -> None:
         if result_backend is None:
             result_backend = DummyResultBackend()
@@ -327,8 +72,8 @@ class AsyncBroker(ABC):
     @overload
     def task(
         self,
-        task_name: Callable[_FuncParams, _ReturnType],
-    ) -> AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType]:
+        task_name: Callable[FuncParams_, ReturnType_],
+    ) -> AsyncTaskiqDecoratedTask[FuncParams_, ReturnType_]:
         ...
 
     @overload
@@ -337,8 +82,8 @@ class AsyncBroker(ABC):
         task_name: Optional[str] = None,
         **labels: Union[str, int],
     ) -> Callable[
-        [Callable[_FuncParams, _ReturnType]],
-        AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType],
+        [Callable[FuncParams_, ReturnType_]],
+        AsyncTaskiqDecoratedTask[FuncParams_, ReturnType_],
     ]:
         ...
 
@@ -374,12 +119,12 @@ class AsyncBroker(ABC):
             inner_labels: Dict[str, Union[str, int]],
             inner_task_name: Optional[str] = None,
         ) -> Callable[
-            [Callable[_FuncParams, _ReturnType]],
-            AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType],
+            [Callable[FuncParams_, ReturnType_]],
+            AsyncTaskiqDecoratedTask[FuncParams_, ReturnType_],
         ]:
             def inner(
-                func: Callable[_FuncParams, _ReturnType],
-            ) -> AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType]:
+                func: Callable[FuncParams_, ReturnType_],
+            ) -> AsyncTaskiqDecoratedTask[FuncParams_, ReturnType_]:
                 nonlocal inner_task_name  # noqa: WPS420
                 if inner_task_name is None:
                     inner_task_name = (  # noqa: WPS442
