+++ taskiq/types_helpers.py
@@ -1,0 +1,7 @@
+from typing import TypeVar
+
+from typing_extensions import ParamSpec
+
+T_ = TypeVar("T_")  # noqa: WPS111, WPS120
+FuncParams_ = ParamSpec("FuncParams_")  # noqa: WPS120
+ReturnType_ = TypeVar("ReturnType_")  # noqa: WPS120
